Move listener middleware into src
